Exported mappingRow content from tableRow
